[REFACTOR] 기존 ErrorCode를 더 범용적인 ServiceCode로 변경
변경: ErroCode -> ServiceCode
응답 성공에 대한 서비스코드도 가지고 있으므로 포괄적 용어 사용
